[chore] Updated with more patterns
Commits from the last 6 months of work & Rosalind Practice.

Should reflect an API, practice with nswag to generate an api client, a simple interface to test functionality, a bit of string manipulation, etc.

Some messy bits. The nswag was generated with the swagger.json provided by using OpenApi, not the clients.nswag. Had trouble getting all of that functional, so more time needed. The clients are generated with a FileResponse, not a List of Shapes as well, so need to figure out how to get that Auto-gen more aligned.

The API uses Giraffe and ASPNetCore, but only the aspnet controllers are generated for the client. Need to figure out how to include the Giraffe endpoints.